+++ NEWS.md
@@ -1,3 +1,8 @@
+# sfo 0.1.2
+
+* Updated the passengers and landing datasets - up to Dec 2022
+* Updated the package documentation and examples
+
 # sfo 0.1.1
 
 * Updated the passengers data - up to Dec 2020 (including)